chore(deps): bump webpack from 5.98.0 to ^5.99.7
