+++ apps/gauzy/package.json
@@ -19,8 +19,8 @@
 	},
 	"resolutions": {
 		"uglify-js": "npm:terser",
-		"webpack": "^5.91.0",
-		"webpack-cli": "^4.9.0"
+		"webpack": "^5.99.7",
+		"webpack-cli": "^6.0.1"
 	},
 	"dependencies": {
 		"@akveo/ng2-completer": "^9.0.1",
@@ -196,9 +196,9 @@
 		"karma-jasmine-html-reporter": "^1.5.4",
 		"terser": "^5.30.3",
 		"terser-webpack-plugin": "^5.3.10",
-		"webpack": "^5.91.0",
-		"webpack-cli": "^4.9.1",
-		"webpack-merge": "^5.10.0",
+		"webpack": "^5.99.7",
+		"webpack-cli": "^6.0.1",
+		"webpack-merge": "^6.0.1",
 		"webpack-node-externals": "^3.0.0"
 	}
 }
